pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fronius-ha-dual-mppt"
version = "0.1.0"
description = "Fronius HA Dual MPPT bridge for Home Assistant"
requires-python = ">=3.8"
dependencies = [
    "pysunspec2>=1.0.0",
    "paho-mqtt>=1.6.0",
    "PyYAML>=6.0",
    "hypothesis>=6.0.0",
]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
